napravljen all session sreen i podeseno da se ode na taj ekran

+++ app/src/main/java/com/example/myapplication/data/auth/responses/SetResponse.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.data.auth.responses
 data class SetResponse(
     val id: Integer,
     val setNumber: Integer,
-    val weight: Integer,
+    val weight: Double,
     val reps: Integer,
     val rpe: Double
 )
